requirements.txt
sqlite-utils>=3.36
pyyaml>=6.0
